consolidate functions
